fixed win build
